feat(comment): user entity 연결

+++ src/main/java/xyz/tomorrowlearncamp/outsourcing/domain/comment/service/OwnerWriteCommentService.java
@@ -6,12 +6,15 @@ import jakarta.validation.constraints.Size;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 import org.springframework.validation.annotation.Validated;
 import xyz.tomorrowlearncamp.outsourcing.domain.comment.entity.CommentEntity;
 import xyz.tomorrowlearncamp.outsourcing.domain.comment.repository.CommentRepository;
 import xyz.tomorrowlearncamp.outsourcing.domain.comment.enums.ErrorCommentMessage;
 import xyz.tomorrowlearncamp.outsourcing.domain.review.entity.ReviewEntity;
 import xyz.tomorrowlearncamp.outsourcing.domain.review.service.UserReadReviewService;
+import xyz.tomorrowlearncamp.outsourcing.domain.user.entity.UserEntity;
+import xyz.tomorrowlearncamp.outsourcing.domain.user.service.UserService;
 import xyz.tomorrowlearncamp.outsourcing.global.exception.InvalidRequestException;
 
 @Service
@@ -21,23 +24,31 @@ public class OwnerWriteCommentService {
 
     private final CommentRepository commentRepository;
 
+    private final UserService userService;
+
     private final UserReadReviewService reviewService;
 
     @Transactional
-    public void saveComment(@NotNull @Positive Long reviewId, @NotNull @Size(min = 1, max = 100) String comment) {
+    public void saveComment(Long userId, @NotNull @Positive Long reviewId, @NotNull @Size(min = 1, max = 100) String comment) {
+
+        UserEntity user = userService.getUserEntity(userId);
 
         ReviewEntity review = reviewService.findReview(reviewId);
 
-        CommentEntity commentEntity = new CommentEntity(review, comment);
+        CommentEntity commentEntity = new CommentEntity(user, review, comment);
 
         commentRepository.save(commentEntity);
     }
 
     @Transactional
-    public void deleteComment(@NotNull @Positive Long commentId) {
+    public void deleteComment(Long userId, @NotNull @Positive Long commentId) {
         CommentEntity deleteComment = commentRepository.findById(commentId).orElseThrow(
                 () -> new InvalidRequestException(ErrorCommentMessage.NOT_FOUND_COMMENT.getErrorMessage())
         );
+
+        if (!ObjectUtils.nullSafeEquals(deleteComment.getUser().getId(), userId)) {
+            throw new InvalidRequestException(ErrorCommentMessage.UNAUTHORIZED.getErrorMessage());
+        }
 
         commentRepository.delete(deleteComment);
     }
